refactor(bridge): extract Desktop IPC layer from cli.py monolith
Decouple the Desktop bridge protocol from core job automation:

- ghosthands/bridge/profile_adapter.py: camelCase→snake_case conversion,
  profile defaults normalization (was inline in cli.py)
- ghosthands/bridge/protocol.py: stdin/stdout IPC, cancel command
  listener, review wait loop (was inline in cli.py)
- ghosthands/agent/prompts.py: add build_task_prompt() (was _build_task_prompt
  in cli.py, belongs with other prompt builders)

cli.py reduced from 1020 → 710 lines. No logic changes.
